refactor(dashboard): promote FP cloud token utils to core-protocols-dashboard
Move `createFPToken`, `getFPTokenContext`, `nameFromAuth`, `nickFromClarkClaim`,
and `toProvider` from `dashboard/backend/utils/` into the shared
`core/protocols/dashboard/fp-cloud-token.ts`. Also move `FPTokenContext`
interface from `dashboard/backend/types.ts` into
`core/types/protocols/dashboard/`. Dashboard backend now imports these
from `@fireproof/core-protocols-dashboard` so they can be reused outside
the backend package.
