+++ core/protocols/dashboard/index.ts
@@ -2,3 +2,5 @@ export * from "./msg-api.js";
 export * from "./dashboard-api.js";
 export * from "./token.js";
 export * from "./get-cloud-pubkey-from-env.js";
+
+export * from "./fp-cloud-token.js";
