+++ core/types/protocols/dashboard/index.ts
@@ -4,3 +4,11 @@ export * from "./msg-is.js";
 export * from "./msg-types.js";
 
 export * from "./token.js";
+export interface FPTokenContext {
+  readonly secretToken: string;
+  readonly publicToken: string;
+  readonly issuer: string;
+  readonly audience: string;
+  readonly validFor: number; // seconds
+  readonly extendValidFor: number; // seconds
+}
