+++ dashboard/backend/public/ensure-cloud-token.ts
@@ -1,5 +1,6 @@
 import { EventoHandler, Result, EventoResultType, HandleTriggerCtx } from "@adviser/cement";
 import {
+  FPTokenContext,
   ReqEnsureCloudToken,
   ResEnsureCloudToken,
   validateEnsureCloudToken,
@@ -9,12 +10,13 @@ import { FPCloudClaimSchema } from "@fireproof/core-types-protocols-cloud";
 import { eq, and, count } from "drizzle-orm";
 import { sqlAppIdBinding } from "../sql/app-id-bind.js";
 import { sqlLedgers, sqlLedgerUsers } from "../sql/ledgers.js";
-import { FPApiSQLCtx, FPTokenContext, ReqWithVerifiedAuthUser } from "../types.js";
-import { getFPTokenContext, createFPToken, toProvider, checkAuth, wrapStop } from "../utils/index.js";
+import { FPApiSQLCtx, ReqWithVerifiedAuthUser } from "../types.js";
+import { checkAuth, wrapStop } from "../utils/index.js";
 import { createLedger } from "./create-ledger.js";
 import { ensureUser } from "./ensure-user.js";
 import { listLedgersByUser } from "./list-ledgers-by-user.js";
 import { decodeJwt } from "jose";
+import { getFPTokenContext, createFPToken, toProvider } from "@fireproof/core-protocols-dashboard";
 
 function getAppIdBinding(
   ctx: FPApiSQLCtx,
