+++ dashboard/backend/public/get-cloud-session-token.ts
@@ -1,11 +1,17 @@
 import { EventoHandler, Result, EventoResultType, HandleTriggerCtx } from "@adviser/cement";
-import { ReqCloudSessionToken, ResCloudSessionToken, validateCloudSessionToken } from "@fireproof/core-types-protocols-dashboard";
+import {
+  FPTokenContext,
+  ReqCloudSessionToken,
+  ResCloudSessionToken,
+  validateCloudSessionToken,
+} from "@fireproof/core-types-protocols-dashboard";
 import { FPCloudClaim } from "@fireproof/core-types-protocols-cloud";
-import { getFPTokenContext, createFPToken, toProvider, checkAuth, wrapStop } from "../utils/index.js";
-import { FPApiSQLCtx, FPTokenContext, ReqWithVerifiedAuthUser } from "../types.js";
+import { checkAuth, wrapStop } from "../utils/index.js";
+import { FPApiSQLCtx, ReqWithVerifiedAuthUser } from "../types.js";
 import { listTenantsByUser } from "./list-tenants-by-user.js";
 import { listLedgersByUser } from "./list-ledgers-by-user.js";
 import { addTokenByResultId } from "../internal/add-token-by-result-id.js";
+import { getFPTokenContext, createFPToken, toProvider } from "@fireproof/core-protocols-dashboard";
 
 async function getCloudSessionToken(
   ctx: FPApiSQLCtx,
